Add cookie based auth to requests, lock method
This commit adds a basic authorization method to all
requests using a cookie stored by the client (so similar
to what is done by Bitcoin RPC). This cookie is sent
in a custom header 'JMCookie'.
Additionally, preflight CORS requests are handled for all
routes generically.
Also more comments are added to the current main functions.
Also a /lock route is added to allow the user to shut down
the wallet service and close the wallet.
Still very rudimentary of course.